+++ CHANGES.md
@@ -9,6 +9,7 @@ Changes in 4.4-beta
 - Leave Delete button AFTER the Save one so info are saved on Enter keypress in Schools.php
 - JS Adjust Side.php submenu bottom offset when footer menu is on top in warehouse.js
 - Do not check allowed tables, sanitize table name instead in Fields.fnc.php
+- Fix extra link when FULL_NAME overridden in Search.inc.php
 
 Changes in 4.3.2
 ----------------